Restore image_histogram2d's output dtype; add CHANGELOG entry
The float32-centers fix (previous commit on this branch) correctly
resolved the large-n_bins float16 bin-collapse bug, but promoted the
function's public hist/pdf return values to float32 too when the input
was float16/bfloat16 -- a real regression flagged in review. Cast hist
back to image.dtype once, at the return boundary, right after the
float32 accumulation completes (so the SUM still happens at full
precision; only the final rounding changes). pdf inherits the same
dtype naturally since it's computed from hist.

Added test_auto_centers_preserve_input_dtype, parametrized over
float16/bfloat16 and return_pdf=True/False (4 cases) per the review's
explicit request for a dtype assertion. Verified RED/GREEN for real:
stashed the histogram.py fix and confirmed all 4 cases fail with
"assert torch.float32 == torch.float16" (exactly the regression being
fixed), then confirmed all 4 pass with the fix restored.

Also fixes a code comment that had actively wrong reasoning: it claimed
"no output-dtype contract to preserve" for centers, missing that
centers' dtype propagates forward to the return values through
ordinary PyTorch type promotion -- which is exactly the bug.

Full tests/enhance/test_histogram.py: 59 passed. ruff check + ruff
format --check clean (v0.16.5 vs. repo's pinned v0.16.4, negligible
patch gap).

+++ kornia/enhance/histogram.py
@@ -240,11 +240,12 @@ def image_histogram2d(
         # once n_bins is large enough: torch.arange(4096, dtype=torch.float16) has only 3073
         # distinct values, not 4096 (confirmed directly), which collapsed centers[i] == centers[j]
         # for i != j and gave those bins numerically identical (redundant) histogram output.
-        # Deliberately NOT cast back to image.dtype: centers is an internal working buffer, not
-        # a returned value, so there is no output-dtype contract to preserve here -- doing so
-        # would re-round every value straight back down to float16 and reproduce the exact same
-        # collapse this fix exists to prevent. `image.unsqueeze(0) - centers` below promotes
-        # normally through ordinary PyTorch type promotion instead.
+        # `centers` stays an internal float32 working buffer for the rest of this function --
+        # `image.unsqueeze(0) - centers` below promotes through ordinary PyTorch type promotion,
+        # so u/kernel_values/hist/pdf all become float32 too unless explicitly cast back down.
+        # That cast happens once, right before `hist` is returned (see below) -- NOT here, which
+        # would re-round centers straight back down to float16 and reproduce the exact collapse
+        # this fix exists to prevent.
         centers = min + bandwidth * (torch.arange(n_bins, device=image.device, dtype=torch.float32) + 0.5)
     centers = centers.reshape(-1, 1, 1, 1, 1)
 
@@ -265,6 +266,12 @@ def image_histogram2d(
         raise ValueError(f"Kernel must be 'triangular', 'gaussian', 'uniform' or 'epanechnikov'. Got {kernel}.")
 
     hist = torch.sum(kernel_values, dim=(-2, -1)).permute(1, 2, 0)
+    # Restore the caller's input dtype here, once, at the public return boundary -- everything
+    # upstream of this point (centers, u, kernel_values) deliberately ran at float32 regardless
+    # of image.dtype (see the `centers is None` branch above); before this cast, a float16/
+    # bfloat16 input silently came back as a float32 hist/pdf, a real behavior change this
+    # dtype-precision fix must not itself introduce.
+    hist = hist.to(image.dtype)
 
     if return_pdf:
         normalization = torch.sum(hist, dim=-1, keepdim=True) + eps
